requirements.txt
attrs==25.1.0
certifi==2025.1.31
charset-normalizer~=3.3.2
h11==0.14.0
idna~=3.7
iniconfig==2.0.0
Jinja2~=3.1.4
MarkupSafe~=2.1.5
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.4
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3~=2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0

pip~=25.0
click~=8.1.7
websockets~=15.0
Werkzeug~=3.0.3
undetected-chromedriver~=3.5.5